requirements.txt
fastapi
uvicorn
virtualenv
tensorflow
pillow